move 'analyze reports list' under  'analyze info'
