+++ src/blitzreplays/replays/analyze_info.py
@@ -11,6 +11,7 @@ import tomlkit
 from pyutils import AsyncTyper
 
 from .models_fields import Fields, ReportField
+from .models_reports import Reports
 from .args import read_param_list, EnumTeamFilter, EnumGroupFilter
 
 
@@ -110,6 +111,59 @@ def info_replay():
     for field in sorted(ReportField._player_fields):
         typer.echo(f"\t{field}")
     typer.echo()
+
+
+@app.command("reports")
+def info_reports(ctx: Context):
+    """
+    List configured reports
+    """
+    reports: Reports = ctx.obj["reports"]
+    reports_param: str | None = ctx.obj["reports_param"]
+    if reports_param is not None:
+        reports = reports.with_config(read_param_list(reports_param))
+    doc: TOMLDocument = tomlkit.document()
+    doc.add("REPORTS", reports.get_toml_report_sets())
+    typer.echo()
+    typer.echo("Configured options for --reports REPORT_SET")
+    typer.echo()
+    typer.echo(tomlkit.dumps(doc))
+    typer.echo()
+    doc = tomlkit.document()
+    doc.add("REPORT", reports.get_toml())
+    typer.echo("Configured reports:")
+    typer.echo()
+    typer.echo(tomlkit.dumps(doc))
+
+
+def read_analyze_reports(config: TOMLDocument) -> Result[Reports, str]:
+    """
+    read REPORT config from analyze TOML config
+    """
+    try:
+        toml_item: TOMLItem | None = None
+        report_store = Reports()
+
+        if "REPORTS" in config and isinstance(
+            toml_item := config.item("REPORTS"), TOMLTable
+        ):
+            for key, report_set in toml_item.items():
+                report_store.add_report_set(key, report_set)
+        else:
+            debug("'REPORTS' is not defined in analyze_config")
+
+        if "REPORT" in config and isinstance(
+            toml_item := config.item("REPORT"), TOMLTable
+        ):
+            for key, rpt in toml_item.items():
+                report_store.add(key=key, **rpt)
+        else:
+            debug("'REPORT' is not defined in analyze_config")
+
+        return Ok(report_store)
+
+    except Exception as err:
+        return Err(str(err))
 
 
 # @app.async_command("add")
